# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Feloxi API",
    "description": "Celery task monitoring and management API",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/workers": {
      "get": {
        "tags": [
          "workers"
        ],
        "operationId": "list_workers",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workers/{worker_id}": {
      "get": {
        "tags": [
          "workers"
        ],
        "operationId": "get_worker",
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "description": "Worker ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workers/{worker_id}/shutdown": {
      "post": {
        "tags": [
          "workers"
        ],
        "operationId": "shutdown_worker",
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "description": "Worker ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CommandResponse": {
        "type": "object",
        "description": "Response for task retry/revoke and worker shutdown commands.",
        "required": [
          "status"
        ],
        "properties": {
          "command_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "task_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "WorkerDetailResponse": {
        "type": "object",
        "description": "Single worker detail with current state and recent events.",
        "required": [
          "worker_id",
          "recent_events"
        ],
        "properties": {
          "current_state": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WorkerState"
              }
            ]
          },
          "recent_events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkerEventRow"
            }
          },
          "worker_id": {
            "type": "string"
          }
        }
      },
      "WorkerEventRow": {
        "type": "object",
        "required": [
          "tenant_id",
          "event_id",
          "worker_id",
          "hostname",
          "event_type",
          "timestamp",
          "active_tasks",
          "processed",
          "load_avg",
          "cpu_percent",
          "memory_mb",
          "pool_size",
          "pool_type",
          "sw_ident",
          "sw_ver",
          "agent_id"
        ],
        "properties": {
          "active_tasks": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "agent_id": {
            "type": "string",
            "format": "uuid"
          },
          "cpu_percent": {
            "type": "number",
            "format": "double"
          },
          "event_id": {
            "type": "string",
            "format": "uuid"
          },
          "event_type": {
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "load_avg": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          },
          "memory_mb": {
            "type": "number",
            "format": "double"
          },
          "pool_size": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pool_type": {
            "type": "string"
          },
          "processed": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "sw_ident": {
            "type": "string"
          },
          "sw_ver": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          },
          "worker_id": {
            "type": "string"
          }
        }
      },
...
      "WorkerListResponse": {
        "type": "object",
        "description": "Worker list with online set, ClickHouse events, and Redis states.",
        "required": [
          "online_workers",
          "worker_events",
          "worker_states"
        ],
        "properties": {
          "online_workers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "worker_events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkerEventRow"
            }
          },
          "worker_states": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkerState"
            }
          }
        }
      },
      "WorkerState": {
        "type": "object",
        "description": "Cached worker state from Redis (set by broker pipeline on heartbeat events).",
        "required": [
          "worker_id",
          "hostname",
          "status",
          "active_tasks",
          "processed",
          "load_avg",
          "cpu_percent",
          "memory_mb",
          "pool_size",
          "pool_type",
          "sw_ident",
          "sw_ver"
        ],
        "properties": {
          "active_tasks": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "cpu_percent": {
            "type": "number",
            "format": "double"
          },
          "hostname": {
            "type": "string"
          },
          "load_avg": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          },
          "memory_mb": {
            "type": "number",
            "format": "double"
          },
          "pool_size": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pool_type": {
            "type": "string"
          },
          "processed": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "status": {
            "type": "string"
          },
          "sw_ident": {
            "type": "string"
          },
          "sw_ver": {
            "type": "string"
          },
          "worker_id": {
            "type": "string"
          }
        }
      }
    }
  },
...
}